+++ src/sync/converters.ts
@@ -1,7 +1,6 @@
 import {
   setRawData,
   parseTimePropertyValue,
-  Relationship,
   RelationshipClass,
 } from '@jupiterone/integration-sdk-core';
 
@@ -36,7 +35,7 @@ import {
   AccountType,
   PullsListResponseItem,
   PullsListCommitsResponseItem,
-  RepoTeamRelationship,
+  RepoAllowRelationship,
 } from '../types';
 import { Approval } from '../approval/collectCommitsForPR';
 import {
@@ -322,7 +321,30 @@ export function createRepoAllowsTeamRelationship(
   repo: RepoEntity,
   team: TeamEntity,
   permission: string,
-): RepoTeamRelationship {
+): RepoAllowRelationship {
+  let admin = false;
+  let maintain = false;
+  let push = false;
+  let triage = false;
+  const pull = true;
+  if (permission === 'TRIAGE') {
+    triage = true;
+  }
+  if (permission === 'WRITE') {
+    triage = true;
+    push = true;
+  }
+  if (permission === 'MAINTAIN') {
+    triage = true;
+    push = true;
+    maintain = true;
+  }
+  if (permission === 'ADMIN') {
+    triage = true;
+    push = true;
+    maintain = true;
+    admin = true;
+  }
   return {
     _key: `${repo._key}|allows|${team._key}`,
     _class: RelationshipClass.ALLOWS,
@@ -330,7 +352,12 @@ export function createRepoAllowsTeamRelationship(
     _fromEntityKey: repo._key,
     _toEntityKey: team._key,
     displayName: RelationshipClass.ALLOWS,
-    permission: permission,
+    permissionType: permission,
+    adminPermission: admin,
+    maintainPermission: maintain,
+    pushPermission: push,
+    triagePermission: triage,
+    pullPermission: pull,
   };
 }
 
@@ -338,7 +365,20 @@ export function createRepoAllowsUserRelationship(
   repo: RepoEntity,
   user: UserEntity,
   permissions?: CollaboratorPermissions,
-): Relationship {
+): RepoAllowRelationship {
+  let permissionType = 'READ';
+  if (permissions?.triage) {
+    permissionType = 'TRIAGE';
+  }
+  if (permissions?.push) {
+    permissionType = 'WRITE';
+  }
+  if (permissions?.maintain) {
+    permissionType = 'MAINTAIN';
+  }
+  if (permissions?.admin) {
+    permissionType = 'ADMIN';
+  }
   return {
     _key: `${repo._key}|allows|${user._key}`,
     _class: RelationshipClass.ALLOWS,
@@ -346,10 +386,11 @@ export function createRepoAllowsUserRelationship(
     _fromEntityKey: repo._key,
     _toEntityKey: user._key,
     displayName: RelationshipClass.ALLOWS,
-    adminPermission: permissions?.admin,
-    maintainPermission: permissions?.maintain,
-    pushPermission: permissions?.push,
-    triagePermission: permissions?.triage,
-    pullPermission: permissions?.pull,
+    permissionType: permissionType,
+    adminPermission: permissions?.admin || false,
+    maintainPermission: permissions?.maintain || false,
+    pushPermission: permissions?.push || false,
+    triagePermission: permissions?.triage || false,
+    pullPermission: permissions?.pull || false,
   };
 }